Fold src/test-ot-glyph and src/test-ot-color into src/main
